+++ tests/testthat/test_adonis.R
@@ -4,4 +4,7 @@ test_that("adonis function works fine", {
   expect_s3_class(adonis_pq(data_fungi, "Tree_name"), "anova")
   expect_s3_class(adonis_pq(subset_samples(data_fungi, !is.na(data_fungi@sam_data$Time)), "Time*Tree_name"), "anova")
   expect_error(adonis_pq(data_fungi, "Time*Tree_name"))
+  expect_error(adonis_pq(enterotype, "SeqTech*Enterotype"))
+  expect_error(adonis_pq(enterotype, "SecTech"))
+  expect_error(adonis_pq(enterotype, "SeqTech", dist_method = "aitchison"))
 })